==> 2-Integrating.Rmd <==
---
title: "Integrating Data"
author: Alina Kline-Schoder
date: | 
      | Started on 04/19/2021
      | Compiled: `r format(Sys.Date(), "%B %d, %Y")`
output:
  github_document:
    toc: yes
  pdf_document:
    toc: yes
  html_notebook:
    df_print: paged
    highlight: tango
    number_sections: yes
    theme: cosmo
    toc: yes
    toc_float: yes
  html_document:
    df_print: paged
    toc: yes
    toc_float: no
editor_options:
  chunk_output_type: inline
always_allow_html: yes
---
# Integrating all data

```{r message = FALSE, echo = FALSE}
rm(list = ls())
library(Seurat)
require(ggdendro)
require(Rmisc)
library(Matrix)
library(MASS)
library(xtable)
library(Matrix.utils)
library(reshape2)
library(BiocManager)
library(rhdf5)
library(ggplot2)
library(dplyr)
library(tidyr)
library(ggpubr)
library(tidyverse)
library(ggsignif)
library(broom)
library(rstatix)
library(ggpmisc)
library(gridExtra)
library(here)
options(future.globals.maxSize = 4000 * 1024^5)
setwd('F:/Data/Sequencing/SavedData/')
load("1-Demultiplexing.RData")
```
  

# QC, RPCA Integration
```{r Integrating Hashing and Run0}
Data<-list(EA001,EA002,EA003,EA004,EA005)

Data <- lapply(X = Data, FUN = function(x) {
  DefaultAssay(x)<-"RNA"
  x <- subset(x,subset = nCount_RNA > 2000 & nCount_RNA < 15000 & percent.mt < 10 & nFeature_RNA>1000)
  x <- NormalizeData(x)
  x <- FindVariableFeatures(x, selection.method = "vst", nfeatures = 2000)
})

# select features that are repeatedly variable across datasets for integration run PCA on each
# dataset using these features
feat <- SelectIntegrationFeatures(object.list = Data)
Data <- lapply(X = Data, FUN = function(x) {
    x <- ScaleData(x, features = feat, verbose = FALSE)
    x <- RunPCA(x, features = feat, verbose = FALSE)
})

anchors <- FindIntegrationAnchors(object.list = Data, anchor.features = feat, reduction = "rpca")

Integrated <- IntegrateData(anchorset = anchors)
```


```{r Saving}
setwd("F:/Data/Sequencing/SavedData/")
save(Integrated,file = "2-Integrating.RData")
```
